Split chart generation into hooks
